pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ninjatradebuilder"
version = "0.1.0"
description = "Historical CL packet validation scaffold for NinjaTradeBuilder v1."
requires-python = ">=3.11"
dependencies = [
  "google-genai>=1.0,<2",
  "pydantic>=2.8,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
